refactor: member 도메인 구현체 클래스 implement 패키지에서 보관
