+++ momoim-external-api/src/main/java/com/triplem/momoim/api/gathering/service/GatheringManagementService.java
@@ -3,8 +3,8 @@ package com.triplem.momoim.api.gathering.service;
 import com.triplem.momoim.core.domain.gathering.dto.ModifyGathering;
 import com.triplem.momoim.core.domain.gathering.infrastructure.GatheringRepository;
 import com.triplem.momoim.core.domain.gathering.model.Gathering;
-import com.triplem.momoim.core.domain.member.GatheringMemberAppender;
-import com.triplem.momoim.core.domain.member.GatheringMemberRemover;
+import com.triplem.momoim.core.domain.member.implement.GatheringMemberAppender;
+import com.triplem.momoim.core.domain.member.implement.GatheringMemberRemover;
 import com.triplem.momoim.exception.BusinessException;
 import com.triplem.momoim.exception.ExceptionCode;
 import lombok.RequiredArgsConstructor;
